+++ src/components/codepreview.tsx
@@ -1,0 +1,45 @@
+import { useRef, useEffect } from "react";
+
+interface CodePreviewProps {
+    code: string;
+}
+
+const html = `
+<html>
+<head></head>
+<body>
+<div id="root"></div>
+<script>
+window.addEventListener('message',(event) => {
+    try {
+       eval(event.data)
+    } catch(error){
+        const root = document.querySelector('#root');
+        root.innerHTML = '<div style="color:red"><h4>Runtime Error</h4>' + error + '</div>'
+    }
+},false);
+</script>
+</body>
+</html>
+`;
+
+const CodePreviewComponent: React.FC<CodePreviewProps> = ({ code }) => {
+    const iframeRef = useRef<any>();
+
+    useEffect(() => {
+        //resetting the htm block within the iframe
+        iframeRef.current.srcdoc = html;
+        iframeRef.current.contentWindow.postMessage(code, "*"); // postMessage takes in 2 arguments and the 2nd one represents the domain name allowed
+    }, [code]);
+
+    return (
+        <div>
+            <iframe title="mirror"
+                ref={iframeRef}
+                sandbox="allow-scripts"
+                srcDoc={html} />
+        </div>
+    );
+};
+
+export default CodePreviewComponent;
